backend/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
certifi==2023.11.17
click==8.1.7
Deprecated==1.2.14
elastic-transport==8.11.0
elasticsearch==8.11.1
fastapi==0.105.0
fastapi-utils==0.2.1
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
importlib-metadata==6.11.0
loguru==0.7.2
opentelemetry-api==1.21.0
opentelemetry-exporter-prometheus==0.42b0
opentelemetry-instrumentation==0.42b0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
prometheus-client==0.19.0
pydantic==1.10.13
pydantic_core==2.14.5
python-dotenv==1.0.0
schedule==1.2.1
sniffio==1.3.0
SQLAlchemy==1.4.50
starlette==0.27.0
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.24.0.post1
wrapt==1.16.0
zipp==3.17.0
